Map program cards to a different component
